deleted additional method, changed maintest name and parsons test comment
